pull audio logic into separate file
